Flytter CalendarEvent til egen fil, endrer orange til orange 300
