+++ force-app/main/freelanceCalendar/lwc/hot_calendar/hot_calendar.js
@@ -4,6 +4,7 @@ import { loadScript, loadStyle } from 'lightning/platformResourceLoader';
 import getCalendarEvents from '@salesforce/apex/HOT_FullCalendarController.getCalendarEvents';
 import IKONER from '@salesforce/resourceUrl/ikoner';
 import { NavigationMixin } from 'lightning/navigation';
+import { CalendarEvent } from './calendar_event';
 
 export default class LibsFullCalendar extends NavigationMixin(LightningElement) {
     static MILLISECONDS_PER_DAY = 86400000;
@@ -313,62 +314,5 @@ export default class LibsFullCalendar extends NavigationMixin(LightningElement) 
             },
             state: { list: listType, from: 'calendar', id: eventExtendedProps.recordId }
         });
-    }
-}
-
-class CalendarEvent {
-    static RED_200 = '#F68282';
-    static BLUE_200 = '#99C3FF';
-    static GREEN_300 = '#66C786';
-    static ORANGE_200 = '#FFD799';
-
-    static eventTypeProperties = new Map([
-        [
-            'SERVICE_APPOINTMENT',
-            {
-                fontColor: 'black',
-                color: CalendarEvent.BLUE_200,
-                pastFontColor: 'black',
-                pastColor: CalendarEvent.RED_200
-            }
-        ],
-        ['COMPLETED_SERVICE_APPOINTMENT', { fontColor: 'black', color: CalendarEvent.GREEN_300 }],
-        ['OPEN_WAGE_CLAIM', { fontColor: 'black', color: CalendarEvent.ORANGE_200 }]
-    ]);
-
-    recordId;
-    type;
-    title;
-    saNumber;
-    description;
-    start;
-    end;
-    isPast;
-    color;
-    textColor;
-
-    /**
-     *
-     * @param {*} data - Forventer data i form av fra HOT_FullCalendarController.CalendarEvent
-     */
-    constructor(data) {
-        this.recordId = data.id;
-        this.saNumber = data.appointmentNumber ?? '';
-        this.description = data.description;
-        this.title = this.saNumber;
-        this.start = new Date(data.startTime);
-        this.end = new Date(data.endTime);
-        this.type = data.type;
-        this.isPast = this.end <= new Date();
-        const properties = this.getEventProperties(this.type) ?? {
-            upcomingColor: '#90cce8',
-            pastColor: '#90cce8'
-        };
-        this.color = (this.isPast ? properties.pastColor : properties.color) ?? properties.color;
-        this.textColor = (this.isPast ? properties.pastFontColor : properties.fontColor) ?? properties.fontColor;
-    }
-
-    getEventProperties(eventType) {
-        return CalendarEvent.eventTypeProperties.get(eventType);
     }
 }
